Rename and move ChildrenMapper -> NodeAllocator
