+++ src/test/java/org/cqfn/astranaut/core/example/green/SimpleAssignment.java
@@ -30,9 +30,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
+import org.cqfn.astranaut.core.algorithms.NodeAllocator;
 import org.cqfn.astranaut.core.base.Builder;
 import org.cqfn.astranaut.core.base.ChildDescriptor;
-import org.cqfn.astranaut.core.base.ChildrenMapper;
 import org.cqfn.astranaut.core.base.EmptyFragment;
 import org.cqfn.astranaut.core.base.Fragment;
 import org.cqfn.astranaut.core.base.Node;
@@ -40,7 +40,7 @@ import org.cqfn.astranaut.core.base.Type;
 import org.cqfn.astranaut.core.utils.ListUtils;
 
 /**
- * Node that describes the 'SimpleAssignment' type..
+ * Node that describes the 'SimpleAssignment' type.
  *
  * @since 1.1.0
  */
@@ -276,9 +276,9 @@ public final class SimpleAssignment implements Assignment {
         @Override
         public boolean setChildrenList(final List<Node> list) {
             final Node[] mapping = new Node[Constructor.MAX_NODE_COUNT];
-            final ChildrenMapper mapper =
-                new ChildrenMapper(SimpleAssignment.TYPE.getChildTypes());
-            final boolean result = mapper.map(mapping, list);
+            final NodeAllocator allocator =
+                new NodeAllocator(SimpleAssignment.TYPE.getChildTypes());
+            final boolean result = allocator.allocate(mapping, list);
             if (result) {
                 this.left = (AssignableExpression) mapping[Constructor.LEFT_POS];
                 this.right = (Expression) mapping[Constructor.RIGHT_POS];
